+++ src/main/java/com/nerdtic/humanos/repositories/AbscenceRepository.java
@@ -3,5 +3,5 @@ package com.nerdtic.humanos.repositories;
 import com.nerdtic.humanos.entities.Abscence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AbscenceRepository extends JpaRepository<Abscence, Integer> {
+public interface AbscenceRepository extends JpaRepository<Abscence, Long> {
 }
