+++ src/main/java/com/nerdtic/humanos/repositories/PosteRepository.java
@@ -3,5 +3,5 @@ package com.nerdtic.humanos.repositories;
 import com.nerdtic.humanos.entities.Poste;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PosteRepository extends JpaRepository<Poste, Integer> {
+public interface PosteRepository extends JpaRepository<Poste, Long> {
 }
