+++ src/main/java/com/nerdtic/humanos/repositories/RoleUtilisateurRepository.java
@@ -3,5 +3,6 @@ package com.nerdtic.humanos.repositories;
 import com.nerdtic.humanos.entities.RoleUtilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleUtilisateurRepository extends JpaRepository<RoleUtilisateur, Integer> {
+public interface RoleUtilisateurRepository extends JpaRepository<RoleUtilisateur, Long> {
+
 }
