+++ src/main/java/com/nerdtic/humanos/repositories/TypeContratRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TypeContratRepository extends JpaRepository<TypeContrat, Integer> {
+public interface TypeContratRepository extends JpaRepository<TypeContrat, Long> {
     List<TypeContrat> findAllByNameContaining(String name);
 }
